+++ tests/testthat/test-dfm_group.R
@@ -26,15 +26,11 @@ test_that("dfm_group works with empty documents", {
         matrix(c(1, 0, 2, 0, 3, 0, 1, 0), nrow = 2, 
                dimnames = list(c("doc1", "doc2"), c("a", "b", "c", "d")))
     )
-})
-
-test_that("dfm_group works with docvars", {
-    mycorpus <- corpus(c("a a b", "a b c c", "a c d d", "a c c d"),
-                       docvars = data.frame(grp = c(1, 1, 2, 2)))
-    mydfm <- dfm(tokens(mycorpus))
-    expect_equal(
-        colSums(dfm_group(mydfm, groups = grp)),
-        colSums(mydfm)
+    
+    expect_equivalent(
+        as.matrix(dfm_group(dfm_remove(dfmt, "*"), c(1, 1, 2))),
+        matrix(numeric(), nrow = 2, 
+               dimnames = list(c("doc1", "doc2"), NULL))
     )
 })